expense_audit_env/pyproject.toml
[project]
name = "expense-audit-env"
version = "0.1.0"
description = "OpenEnv environment for corporate expense report auditing — trains AI agents to review expenses for policy compliance."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "BSD-3-Clause"}

dependencies = [
    "openenv-core>=0.2.0",
    "pydantic>=2.0.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "httpx>=0.24.0",
]

[project.scripts]
server = "server.app:main"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"
